Combine _util.js and _utils.js
