+++ src/pages/dashboard/charts/_AreaChart.js
@@ -6,7 +6,7 @@ import { scaleLinear, scaleTime } from 'd3-scale'
 import { area } from 'd3-shape'
 import React from 'react'
 
-import { formatDate, formatNumber } from '../_util'
+import { formatDate, formatNumber } from '../_utils'
 
 export default function AreaChart({
   annotations = null,
